+++ MoreShipUpgrades/Managers/PatchManager.cs
@@ -61,6 +61,7 @@ namespace MoreShipUpgrades.Managers
             harmony.PatchAll(typeof(DoorLockPatcher));
             harmony.PatchAll(typeof(HangarShipDoorPatcher));
             harmony.PatchAll(typeof(InteractTriggerPatcher));
+            harmony.PatchAll(typeof(QuicksandTriggerPatcher));
             harmony.PatchAll(typeof(SteamValveHazardPatch));
             Plugin.mls.LogInfo("Interactables have been patched");
         }