+++ MoreShipUpgrades/UpgradeComponents/TierUpgrades/Player/OxygenCanisters.cs
@@ -28,7 +28,7 @@ namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades.Player
         }
         public override string GetDisplayInfo(int initialPrice = -1, int maxLevels = -1, int[] incrementalPrices = null)
         {
-            System.Func<int, float> infoFunction = level => UpgradeBus.Instance.PluginConfiguration.REINFORCED_BOOTS_INITIAL_DAMAGE_REDUCTION.Value + level * UpgradeBus.Instance.PluginConfiguration.REINFORCED_BOOTS_INCREMENTAL_DAMAGE_REDUCTION.Value;
+            System.Func<int, float> infoFunction = level => UpgradeBus.Instance.PluginConfiguration.OXYGEN_CANISTERS_INITIAL_OXYGEN_CONSUMPTION_DECREASE.Value + level * UpgradeBus.Instance.PluginConfiguration.REINFORCED_BOOTS_INCREMENTAL_DAMAGE_REDUCTION.Value;
             string infoFormat = "LVL {0} - ${1} - Reduces oxygen consumption rate by {2}%\n";
             return Tools.GenerateInfoForUpgrade(infoFormat, initialPrice, incrementalPrices, infoFunction);
         }